Changed volume type for engine

+++ AutoDealer/AutoDealer.Business/Models/Commands/Car/CarEngineCreateCommand.cs
@@ -5,11 +5,11 @@ namespace AutoDealer.Business.Models.Commands.Car
     public class CarEngineCreateCommand : ICreateCommand
     {
         public string Name { get; }
-        public int Volume { get; }
+        public float Volume { get; }
         public int Power { get; }
         public int TypeId { get; }
 
-        public CarEngineCreateCommand(string name, int volume, int power, int typeId)
+        public CarEngineCreateCommand(string name, float volume, int power, int typeId)
         {
             Name = name;
             Volume = volume;
