+++ AutoDealer/AutoDealer.Business/Models/Responses/Car/CarEngineModel.cs
@@ -3,11 +3,11 @@
     public class CarEngineModel : BaseModel
     {
         public string Name { get; }
-        public int Volume { get; }
+        public float Volume { get; }
         public int Power { get; }
         public CarEngineTypeModel Type { get; }
 
-        public CarEngineModel(int id, string name, int volume, int power, CarEngineTypeModel type) : base(id)
+        public CarEngineModel(int id, string name, float volume, int power, CarEngineTypeModel type) : base(id)
         {
             Name = name;
             Volume = volume;
