+++ AutoDealer/AutoDealer.Data/Models/Car/CarEngine.cs
@@ -5,7 +5,7 @@ namespace AutoDealer.Data.Models.Car
     public class CarEngine : BaseModel
     {
         public string Name { get; set; }
-        public int Volume { get; set; }
+        public float Volume { get; set; }
         public int Power { get; set; }
         public int TypeId { get; set; }
         public CarEngineType Type { get; set; }
