+++ AutoDealer/AutoDealer.Web/ViewModels/Request/Car/CarEngineCreateViewModel.cs
@@ -3,7 +3,7 @@
     public class CarEngineCreateViewModel
     {
         public string Name { get; set; }
-        public int Volume { get; set; }
+        public float Volume { get; set; }
         public int Power { get; set; }
         public int TypeId { get; set; }
     }
