+++ AutoDealer/AutoDealer.Web/ViewModels/Request/Car/CarEngineUpdateViewModel.cs
@@ -5,7 +5,7 @@ namespace AutoDealer.Web.ViewModels.Request.Car
     public class CarEngineUpdateViewModel : BaseViewModel
     {
         public string Name { get; set; }
-        public int Volume { get; set; }
+        public float Volume { get; set; }
         public int Power { get; set; }
         public int TypeId { get; set; }
     }
